agents/adk-python/requirements.txt
fastapi
uvicorn[standard]
httpx
google-cloud-aiplatform

google-genai>=1.32.0
